Merge Develop into Master
Agregados los endpoints para comunicación co la base de datos, con algunas protecciones y con métodos ayudantes para la realización mas "organizada" de las consultas.
Se cambia la cadena de conexión a una clase "Constants", para que solo deba cambiarse una vez al cambiar de base de datos.

+++ Repository/ProductoHandler.cs
@@ -5,40 +5,6 @@ namespace SistemaGestionWebApi_EnzoDonadel.Repository
 {
     internal class ProductoHandler
     {
-        //Traer Productos (recibe un id de usuario y, devuelve una lista con todos los productos cargado por ese usuario)
-        public static List<Producto> getProductByUserId(long userIdToSearch)
-        {
-            List<Producto> products = new List<Producto>();
-            using (SqlConnection SqlDbConnection = new SqlConnection(Constants.connectionString))
-            {
-                using (SqlCommand SqlDbQuery = new SqlCommand("SELECT * FROM Producto WHERE IdUsuario =@parameterToSearch", SqlDbConnection))
-                {
-                    SqlParameter ParameterID = new SqlParameter("parameterToSearch", System.Data.SqlDbType.BigInt);
-                    ParameterID.Value = userIdToSearch;
-                    SqlDbQuery.Parameters.Add(ParameterID);
-                    SqlDbConnection.Open();
-                    using (SqlDataReader DataReader = SqlDbQuery.ExecuteReader())
-                    {
-                        if (DataReader.HasRows)
-                        {
-                            while (DataReader.Read())
-                            {
-                                Producto temp = new Producto();
-                                temp.Id = DataReader.GetInt64(0);
-                                temp.Descripcion = DataReader.GetString(1);
-                                temp.Costo = DataReader.GetDecimal(2);
-                                temp.PrecioVenta = DataReader.GetDecimal(3);
-                                temp.Stock = DataReader.GetInt32(4);
-                                temp.IdUsuario = DataReader.GetInt64(5);
-                                products.Add(temp);
-                            }
-                        }
-                    }
-                    SqlDbConnection.Close();
-                }
-            }
-            return products;
-        }
         public static Producto getProductById(long IdToSearch)
         {
             Producto product = new Producto();
@@ -56,7 +22,7 @@ namespace SistemaGestionWebApi_EnzoDonadel.Repository
                         {
                             DataReader.Read();
                             product.Id = DataReader.GetInt64(0);
-                            product.Descripcion = DataReader.GetString(1);
+                            product.Descripciones = DataReader.GetString(1);
                             product.Costo = DataReader.GetDecimal(2);
                             product.PrecioVenta = DataReader.GetDecimal(3);
                             product.Stock = DataReader.GetInt32(4);
@@ -91,34 +57,74 @@ namespace SistemaGestionWebApi_EnzoDonadel.Repository
             }
             return products;
         }
-        public static void AddProducto(Producto productToAdd)
+        #region Metodos Proyecto Final
+        //Traer Productos (recibe un id de usuario y, devuelve una lista con todos los productos cargado por ese usuario)
+        public static List<Producto> getProductsByUserId(long userIdToSearch)
         {
-            int AffectedRegisters;
-            string query = "INSERT INTO Producto " +
-                "(Descripciones, Costo, PrecioVenta, Stock, IdUsuario) " +
-                "VALUES " +
-                "(@DescriptionToADD, @CostoToADD, @PrecioVentaToADD, @StockToADD, @IdUsuarioToADD)";
+            List<Producto> products = new List<Producto>();
+            string query = "SELECT * FROM Producto WHERE IdUsuario =@parameterToSearch";
             using (SqlConnection SqlDbConnection = new SqlConnection(Constants.connectionString))
             {
                 using (SqlCommand SqlDbQuery = new SqlCommand(query, SqlDbConnection))
                 {
-                    Producto original = getProductById(productToAdd.Id);
-                    SqlDbQuery.Parameters.AddWithValue("@DescriptionToADD", productToAdd.Descripcion);
+                    SqlDbQuery.Parameters.AddWithValue("@parameterToSearch", userIdToSearch);
+                    SqlDbConnection.Open();
+                    using (SqlDataReader DataReader = SqlDbQuery.ExecuteReader())
+                    {
+                        if (DataReader.HasRows)
+                        {
+                            while (DataReader.Read())
+                            {
+                                Producto temp = new Producto();
+                                temp.Id = DataReader.GetInt64(0);
+                                temp.Descripciones = DataReader.GetString(1);
+                                temp.Costo = DataReader.GetDecimal(2);
+                                temp.PrecioVenta = DataReader.GetDecimal(3);
+                                temp.Stock = DataReader.GetInt32(4);
+                                temp.IdUsuario = DataReader.GetInt64(5);
+                                products.Add(temp);
+                            }
+                        }
+                    }
+                    SqlDbConnection.Close();
+                }
+            }
+            return products;
+        }
+        public static bool AddProducto(Producto productToAdd)
+        {
+            bool result = false;
+            string query = "INSERT INTO Producto " +
+                                "(Descripciones, Costo, PrecioVenta, Stock, IdUsuario) " +
+                            "VALUES " +
+                                "(@DescriptionToADD, @CostoToADD, @PrecioVentaToADD, @StockToADD, @IdUsuarioToADD)";
+            using (SqlConnection SqlDbConnection = new SqlConnection(Constants.connectionString))
+            {
+                using (SqlCommand SqlDbQuery = new SqlCommand(query, SqlDbConnection))
+                {
+                    SqlDbQuery.Parameters.AddWithValue("@DescriptionToADD", productToAdd.Descripciones);
                     SqlDbQuery.Parameters.AddWithValue("@CostoToADD", productToAdd.Costo);
                     SqlDbQuery.Parameters.AddWithValue("@PrecioVentaToADD", productToAdd.PrecioVenta);
                     SqlDbQuery.Parameters.AddWithValue("@StockToADD", productToAdd.Stock);
                     SqlDbQuery.Parameters.AddWithValue("@IdUsuarioToADD", productToAdd.IdUsuario);
                     SqlDbConnection.Open();
-                    AffectedRegisters = SqlDbQuery.ExecuteNonQuery();
+                    if (SqlDbQuery.ExecuteNonQuery() == 1)
+                    {
+                        result = true;
+                    }
                     SqlDbConnection.Close();
                 }
             }
+            return result;
         }
-        public static void DeleteProduct(long idToDelete)
+        public static bool DeleteProduct(long idToDelete)
         {
-            int AffectedRegisters;
+            bool result = false;
             //Previamente se deben eliminar todos los productos vendidos del producto en cuestion.
-
+            if (!ProductoVendidoHandler.DeleteProductoVendidoByProductID(idToDelete))
+            {
+                return false;
+            }
             string query = "Delete FROM Producto " +
                             "WHERE " +
                                 "Id = @idParameter";
@@ -126,16 +132,18 @@ namespace SistemaGestionWebApi_EnzoDonadel.Repository
             {
                 using (SqlCommand SqlDbQuery = new SqlCommand(query, SqlDbConnection))
                 {
-                    SqlDbQuery.Parameters.AddWithValue("@IdParameter", idToDelete);
+                    SqlDbQuery.Parameters.AddWithValue("@idParameter", idToDelete);
                     SqlDbConnection.Open();
-                    AffectedRegisters = SqlDbQuery.ExecuteNonQuery();
+                    SqlDbQuery.ExecuteNonQuery();
                     SqlDbConnection.Close();
+                    result = true;
                 }
             }
+            return result;
         }
-        public static void UpdateProducto(Producto DataToUpdate)
+        public static bool UpdateProducto(Producto DataToUpdate)
         {
-            int AffectedRegisters;
+            bool result = false;
             string query = "UPDATE Producto " +
                             "SET " +
                                 "Descripciones = @descriptionToChange, " +
@@ -149,21 +157,28 @@ namespace SistemaGestionWebApi_EnzoDonadel.Repository
                 using (SqlCommand SqlDbQuery = new SqlCommand(query, SqlDbConnection))
                 {
                     Producto original = getProductById(DataToUpdate.Id);
-                    SqlDbQuery.Parameters.AddWithValue("@descriptionToChange", DataToUpdate.Descripcion);
+                    SqlDbQuery.Parameters.AddWithValue("@descriptionToChange", DataToUpdate.Descripciones);
                     SqlDbQuery.Parameters.AddWithValue("@costoToChange", DataToUpdate.Costo);
                     SqlDbQuery.Parameters.AddWithValue("@precioVentaToChange", DataToUpdate.PrecioVenta);
                     SqlDbQuery.Parameters.AddWithValue("@stockToChange", DataToUpdate.Stock);
                     SqlDbQuery.Parameters.AddWithValue("@idUsuarioToChange", DataToUpdate.IdUsuario);
                     SqlDbQuery.Parameters.AddWithValue("@ProductId", DataToUpdate.Id);
                     SqlDbConnection.Open();
-                    AffectedRegisters = SqlDbQuery.ExecuteNonQuery();
+                    if (SqlDbQuery.ExecuteNonQuery() == 1)
+                    {
+                        result = true;
+                    }
                     SqlDbConnection.Close();
                 }
             }
+            return result;
         }
-        public static void UpdateStockProducto(long idProducto, int StockVendido)
+        #endregion
+        #region Metodos Internos
+        //Necesarios para otros metodos
+        internal static bool UpdateStockProducto(long idProducto, int StockVendido)
         {
-            int AffectedRegisters;
+            bool result = false;
             Producto productSelled = ProductoHandler.getProductById(idProducto);
             productSelled.Stock -= StockVendido;
             string query = "UPDATE Producto " +
@@ -177,10 +192,45 @@ namespace SistemaGestionWebApi_EnzoDonadel.Repository
                     SqlDbQuery.Parameters.AddWithValue("@stockToChange", productSelled.Stock);
                     SqlDbQuery.Parameters.AddWithValue("@ProductId", productSelled.Id);
                     SqlDbConnection.Open();
-                    AffectedRegisters = SqlDbQuery.ExecuteNonQuery();
+                    if (SqlDbQuery.ExecuteNonQuery() == 1)
+                    {
+                        result = true;
+                    }
                     SqlDbConnection.Close();
                 }
             }
+            return result;
         }
+        //Metodo Helper necesario para el funcionamiento de "BorrarUsuario", ya que previamente se deben borrar los productos cargados por ese Usuario.
+        internal static bool DeleteProductsByUser(long idUsuario)
+        {
+            bool result = false;
+            int AffectedRegisters;
+            //Previamente se deben eliminar todos los productos vendidos del producto en cuestion.
+            List<Producto> Productos = ProductoHandler.getProductsByUserId(idUsuario);
+            foreach (Producto producto in Productos)
+            {
+                ProductoVendidoHandler.DeleteProductoVendidoByProductID(producto.Id);
+            }
+            string query = "Delete FROM Producto " +
+                            "WHERE " +
+                                "IdUsuario = @idParameter";
+            using (SqlConnection SqlDbConnection = new SqlConnection(Constants.connectionString))
+            {
+                using (SqlCommand SqlDbQuery = new SqlCommand(query, SqlDbConnection))
+                {
+                    SqlDbQuery.Parameters.AddWithValue("@idParameter", idUsuario);
+                    SqlDbConnection.Open();
+                    AffectedRegisters = SqlDbQuery.ExecuteNonQuery();
+                    if (AffectedRegisters > 0)
+                    {
+                        result = true;
+                    }
+                    SqlDbConnection.Close();
+                }
+            }
+            return result;
+        }
+        #endregion
     }
 }
